Fixed a problem in the motd::knife-status recipe
When it gets the "interval" value from node['chef_client']['interval'],
it was failing to convert the seconds to minutes since the value
returned was actually a string. This fix just ensures that the
value in "interval" is actually an integer, so that we can do
math operations on it.

See: https://github.com/opscode-cookbooks/chef-client/blob/master/attributes/default.rb#L36